App.js comments
Overall works as expected great job. For future, consider adding edit functionality and a search component. Also consider adding the ability for a user to group cards together based on a category. 

This file is great. I like your use of the parent / child components as well as your arrow functions and use of const { } = this.state (this improves code readability).